+++ src/framework/numbers.h
@@ -3,4 +3,4 @@ float clipFloat(float, float, float);
 int distance(int, int, int, int);
 float distanceFloat(float, float, float, float);
 int directionTo(int, int, int, int);
-void velocity(float[2], int, float);
+void velocity(double[2], int, float);
